Relocated config to a new config package.
