+++ internal/config/config_test.go
@@ -1,8 +1,12 @@
-package main
+package config_test
 
 import (
 	"testing"
+
+	"github.com/0x4d31/galah/internal/config"
 )
+
+const configPath = "../../config/config.yaml"
 
 func TestLoadConfig(t *testing.T) {
 	tests := []struct {
@@ -11,12 +15,12 @@ func TestLoadConfig(t *testing.T) {
 		wantErr    bool
 	}{
 		{
-			name:       "Load valid configuration",
-			configPath: "config.yaml",
+			name:       "validConfiguration",
+			configPath: configPath,
 			wantErr:    false,
 		},
 		{
-			name:       "Load non-existent configuration",
+			name:       "nonExistentConfiguration",
 			configPath: "path_to_non_existent_config.yaml",
 			wantErr:    true,
 		},
@@ -24,7 +28,7 @@ func TestLoadConfig(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			_, err := LoadConfig(tt.configPath)
+			_, err := config.LoadConfig(tt.configPath)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("LoadConfig() error = %v, wantErr %v", err, tt.wantErr)
 			}
